GamesSite/requirements.txt
Django==1.7.3
Pillow==2.7.0
PyJWT==0.4.1
Twisted==15.0.0
Watson==0.1.05
Whoosh==2.6.0
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-bootstrap3==5.1.1
django-haystack==2.3.1
django-settings==1.3-11
django-social-auth==0.7.28
django-tastypie==0.12.1
django-toolbelt==0.0.1
django-watson==1.1.5
gunicorn==19.1.1
httplib2==0.9
oauth2==1.5.211
oauthlib==0.7.2
pinder==1.0.1
psycopg2==2.5.4
python-dateutil==2.4.0
python-mimeparse==0.1.4
python-openid==2.2.5
python-ptrace==0.8.1
python-social-auth==0.2.1
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
static3==0.5.1
wokkel==0.7.0
wsgiref==0.1.2
zope.interface==4.1.2
